moved dropdown and dropdown styling to own component folder
